Convert web input to general graphical input
